task 2 updated

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -7,7 +7,7 @@
  * @action:  is a pointer to the function you need to use
  * Return: void
  */
-void array_iterator(int *array, size_t size, void (*action)(int))
+void array_iterator(int *array, unsigned int size, void (*action)(int))
 {
 	unsigned int i;
